requirements.txt
aiogram>=3.12.0
aiosqlite>=0.19.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
Pillow>=10.2.0
qrcode>=7.4.2
